+++ AutoReservation.GUI/ViewModels/ReservationViewModel.cs
@@ -1,10 +1,7 @@
 ﻿using AutoReservation.Common.DataTransferObjects;
-using AutoReservation.Common.Interfaces;
-using AutoReservation.GUI.Commands;
 using System;
 using System.Collections.ObjectModel;
 using System.Linq;
-using System.ServiceModel;
 using System.Windows;
 
 namespace AutoReservation.GUI.ViewModels
@@ -13,7 +10,7 @@ namespace AutoReservation.GUI.ViewModels
     {
         public ReservationViewModel()
         {
-            _list = new ObservableCollection<ReservationDto>(target.Reservationen);
+            _list = new ObservableCollection<ReservationDto>(target.Reservationen.FindAll(o => o.Bis > DateTime.Now).OrderBy(o => o.Von).ThenBy(o => o.Bis));
         }
 
         public override void Add()
@@ -29,25 +26,12 @@ namespace AutoReservation.GUI.ViewModels
             return true;
         }
 
-        public override void Modify()
-        {
-            Window DetailsWindow = new ReservationDetailsWindow();
-            DetailsWindow.DataContext = new ReservationDetailsVM(DetailsWindow, Selected);
-            DetailsWindow.Show();
-        }
-
-        public override bool CanModify()
-        {
-            if (Selected == null)
-            {
-                return false;
-            }
-            return true;
-        }
-
         public override void Delete()
         {
-            target.DeleteReservation(Selected);
+            if (MessageBox.Show("Reservation wirklich löschen?", "Bestätigen", MessageBoxButton.YesNo, MessageBoxImage.Warning) == MessageBoxResult.Yes)
+            {
+                target.DeleteReservation(Selected);
+            }
         }
 
         public override bool CanDelete()
@@ -57,6 +41,18 @@ namespace AutoReservation.GUI.ViewModels
                 return false;
             }
             return true;
+        }
+        
+        public override void Modify()
+        {
+            // Not used in this View
+            throw new System.NotImplementedException();
+        }
+        
+        public override bool CanModify()
+        {
+            // Not used in this View
+            throw new System.NotImplementedException();
         }
     }
 }
